refactor: organize components/models/stores/hooks to different folders/files
