+++ src/pages/main/models/tradeModels.ts
@@ -1,0 +1,5 @@
+export enum SlippageStep {
+	Step1, // 0.1%
+	Step2, // 0.5%
+	Step3, // 1.0%
+}
